io: move Sha256Digest from osmet/io_helpers
Those bits are currently only used in osmet, but soon won't be. There's
nothing osmet-specific about the code otherwise.
