+++ src/io.rs
@@ -14,8 +14,14 @@
 
 use anyhow::{bail, ensure, Context, Error, Result};
 use flate2::read::GzDecoder;
+use openssl::hash::{Hasher, MessageDigest};
 use openssl::sha;
+use serde::{Deserialize, Serialize};
+use std::convert::{TryFrom, TryInto};
+use std::fs::OpenOptions;
 use std::io::{self, BufRead, ErrorKind, Read, Write};
+use std::os::unix::io::AsRawFd;
+use std::path::Path;
 use std::result;
 use std::str::FromStr;
 use xz2::read::XzDecoder;
@@ -251,6 +257,61 @@ impl<R: Read> Read for LimitReader<R> {
             .checked_sub(count as u64)
             .expect("read more bytes than allowed");
         Ok(count)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Default)]
+pub struct Sha256Digest(pub [u8; 32]);
+
+impl TryFrom<Hasher> for Sha256Digest {
+    type Error = Error;
+
+    fn try_from(mut hasher: Hasher) -> std::result::Result<Self, Self::Error> {
+        let digest = hasher.finish().context("finishing hash")?;
+        Ok(Sha256Digest(
+            digest.as_ref().try_into().context("converting to SHA256")?,
+        ))
+    }
+}
+
+impl Sha256Digest {
+    /// Calculates the SHA256 of a file.
+    pub fn from_path(path: &Path) -> Result<Self> {
+        let mut f = OpenOptions::new()
+            .read(true)
+            .open(path)
+            .with_context(|| format!("opening {:?}", path))?;
+
+        Self::from_file(&mut f)
+    }
+
+    /// Calculates the SHA256 of an opened file. Note that the underlying file descriptor will have
+    /// `posix_fadvise` called on it to optimize for sequential reading.
+    pub fn from_file(f: &mut std::fs::File) -> Result<Self> {
+        // tell kernel to optimize for sequential reading
+        if unsafe { libc::posix_fadvise(f.as_raw_fd(), 0, 0, libc::POSIX_FADV_SEQUENTIAL) } < 0 {
+            eprintln!(
+                "posix_fadvise(SEQUENTIAL) failed (errno {}) -- ignoring...",
+                nix::errno::errno()
+            );
+        }
+
+        Self::from_reader(f)
+    }
+
+    /// Calculates the SHA256 of a reader.
+    pub fn from_reader(r: &mut impl Read) -> Result<Self> {
+        let mut hasher = Hasher::new(MessageDigest::sha256()).context("creating SHA256 hasher")?;
+        std::io::copy(r, &mut hasher)?;
+        hasher.try_into()
+    }
+
+    pub fn to_hex_string(&self) -> Result<String> {
+        let mut buf: Vec<u8> = Vec::with_capacity(64);
+        for i in 0..32 {
+            write!(buf, "{:02x}", self.0[i])?;
+        }
+        Ok(String::from_utf8(buf)?)
     }
 }
